refactor(search): split observed store tests
